# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/analytics/download-market-data": {
            "get": {
                "description": "Download the market data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DownloadMarketData"
                ],
                "summary": "Download the market data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/analytics/meeting-runners": {
            "get": {
                "description": "Get the meeting runners",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetMeetingRunners"
                ],
                "summary": "Get the meeting runners",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Name",
                        "name": "event_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TodayRunners"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.TodayRunners": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "event_time": {
                    "type": "string"
                },
                "horse_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "selection_id": {
                    "type": "integer"
                }
            }
        }
    }
}